refactor: swap society files
